requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
click==6.7
cryptography==2.3.1
Flask==1.0.2
Flask-MySQL==1.4.0
Flask-RESTful==0.3.6
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.19
PyMySQL==0.9.2
python-dotenv==0.9.1
pytz==2018.5
six==1.11.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
